Fix CLI

+++ aleth-vm-instr/main.cpp
@@ -140,7 +140,7 @@ int main(int argc, char** argv)
     addGaOption("init-program-size", po::value<uint32_t>(), "<n> Set initial program size");
     addGaOption("min-program-size", po::value<uint32_t>(), "<n> Set minimum program size");
     addGaOption("generations-count", po::value<uint32_t>(), "<n> Set numbers of generation to run");
-    addGaOption("elite-ratio", po::value<uint32_t>(), "<x> Set the ratio of elite to keep across generations");
+    addGaOption("elite-ratio", po::value<double>(), "<x> Set the ratio of elite to keep across generations");
     addGaOption("tournament-selection-p", po::value<double>(), "<x> Set probability of picking first for tournament selection");
     addGaOption("tournament-selection-ratio", po::value<double>(), "<x> Set the ratio of samples to use for tournament selection");
